requirements.txt
asgiref==3.6.0
cachetools==5.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.1.1
delegator.py==0.1.1
distlib==0.3.6
Django==4.2.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
filelock==3.8.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.13.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
graphql-core==3.2.3
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==2.6.0
proto-plus==1.22.1
protobuf==4.21.12
psycopg2==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
strawberry-graphql==0.211.1
strawberry-graphql-django==0.21.0
typing_extensions==4.8.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
